+++ server/src/utils/helpers.js
@@ -1,0 +1,7 @@
+// Utility function example for server
+function capitalize(str) {
+  if (typeof str !== 'string') return '';
+  return str.charAt(0).toUpperCase() + str.slice(1);
+}
+
+module.exports = { capitalize };